requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.3.2
pillow==8.3.1
tensorflow-gpu==2.5.0
